+++ lib/common/data/dto/pickup_confirmation_dto.g.dart
@@ -1,0 +1,19 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'pickup_confirmation_dto.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+PickupConfirmationDto _$PickupConfirmationDtoFromJson(
+        Map<String, dynamic> json) =>
+    PickupConfirmationDto(
+      enabled: json['enabled'] as bool?,
+    );
+
+Map<String, dynamic> _$PickupConfirmationDtoToJson(
+        PickupConfirmationDto instance) =>
+    <String, dynamic>{
+      'enabled': instance.enabled,
+    };